feat: move your files to a separate app
