+++ dataworkspace/dataworkspace/apps/core/views.py
@@ -1,6 +1,5 @@
 import logging
 
-from django.conf import settings
 from django.http import (
     HttpResponse,
     HttpResponseRedirect,
@@ -11,12 +10,10 @@ from django.http import (
 from django.shortcuts import render
 from django.urls import reverse
 from django.views.generic import FormView
-from csp.decorators import csp_update
 
 from dataworkspace.apps.core.forms import SupportForm
 from dataworkspace.apps.core.utils import (
     can_access_schema_table,
-    get_s3_prefix,
     table_data,
     table_exists,
     view_exists,
@@ -107,25 +104,3 @@ def table_data_view(request, database, schema, table):
         return HttpResponseNotFound()
     else:
         return table_data(request.user.email, database, schema, table)
-
-
-def file_browser_html_view(request):
-    return (
-        file_browser_html_GET(request)
-        if request.method == 'GET'
-        else HttpResponse(status=405)
-    )
-
-
-@csp_update(
-    CONNECT_SRC=[settings.APPLICATION_ROOT_DOMAIN, "https://s3.eu-west-2.amazonaws.com"]
-)
-def file_browser_html_GET(request):
-    prefix = get_s3_prefix(str(request.user.profile.sso_id))
-
-    return render(
-        request,
-        'files.html',
-        {'prefix': prefix, 'bucket': settings.NOTEBOOKS_BUCKET},
-        status=200,
-    )
