+++ dataworkspace/dataworkspace/settings/base.py
@@ -96,6 +96,7 @@ INSTALLED_APPS = [
     'dataworkspace.apps.api_v1',
     'dataworkspace.apps.eventlog',
     'dataworkspace.apps.request_data',
+    'dataworkspace.apps.your_files',
     'django_extensions',
     'dataworkspace.apps.explorer',
     'dynamic_models',
